chore!: rename Prototyping asmdef to Components
