+++ com.unity.netcode.gameobjects/Components.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4f9f645c043894b81ab326fe4703f4c8
+guid: 8b267eb841a574dc083ac248a95d4443
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
