+++ com.unity.netcode.gameobjects/Components/NetworkAnimator.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Unity.Netcode.Prototyping
+namespace Unity.Netcode.Components
 {
     /// <summary>
     /// A prototype component for syncing animations
