+++ com.unity.netcode.gameobjects/Components/NetworkAnimator.cs.meta
