+++ com.unity.netcode.gameobjects/Components/NetworkTransform.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace Unity.Netcode.Prototyping
+namespace Unity.Netcode.Components
 {
     /// <summary>
     /// A prototype component for syncing transforms
