+++ com.unity.netcode.gameobjects/Components/com.unity.netcode.components.asmdef
@@ -1,0 +1,7 @@
+{
+    "name": "Unity.Netcode.Components",
+    "rootNamespace": "Unity.Netcode.Components",
+    "references": [
+        "Unity.Netcode.Runtime"
+    ]
+}
