+++ com.unity.netcode.gameobjects/Tests/Runtime/com.unity.netcode.runtimetests.asmdef
@@ -4,7 +4,7 @@
     "references": [
         "Unity.Netcode.Runtime",
         "Unity.Netcode.Editor",
-        "Unity.Netcode.Prototyping",
+        "Unity.Netcode.Components",
         "Unity.Collections",
         "UnityEngine.TestRunner",
         "Unity.Multiplayer.MetricTypes",
