+++ testproject/Assets/Tests/Manual/testproject.manualtests.asmdef
@@ -5,6 +5,6 @@
         "TestProject",
         "Unity.Netcode.Runtime",
         "Unity.Netcode.Editor",
-        "Unity.Netcode.Prototyping"
+        "Unity.Netcode.Components"
     ]
 }
